+++ Dockerfile
@@ -4,7 +4,6 @@ LABEL maintainer="info@datascience.ch"
 
 RUN pip install --no-cache-dir --disable-pip-version-check -U pip && \
     pip install --no-cache-dir --disable-pip-version-check pipenv && \
-    pip install --no-cache-dir --disable-pip-version-check yamlpath && \
     groupadd -g 1000 kyaku && \
     useradd -u 1000 -g kyaku kyaku
 
@@ -17,10 +16,6 @@ COPY renku_notebooks /renku-notebooks/renku_notebooks
 COPY resource_schema_migrations /resource_schema_migrations
 # Set up the flask app
 ENV FLASK_APP=/renku-notebooks/renku-notebooks:create_app
-
-# Update version
-COPY helm-chart/renku-notebooks/Chart.yaml /renku-notebooks/
-RUN echo "__version__ = \"$(yaml-get --query .version Chart.yaml)\"" > renku_notebooks/version.py
 
 # Switch to unpriviledged user
 USER kyaku
