+++ renku_notebooks/api/__init__.py
@@ -18,7 +18,6 @@
 """ Notebooks service APIs """
 from flask import Blueprint, Response
 
-from .schemas import VersionResponse
 from .. import config
 from . import notebooks
 
@@ -29,38 +28,6 @@ bp = Blueprint("notebooks_service", __name__)
 def health():
     """Just a health check path."""
     return Response("service running under {}".format(config.SERVICE_PREFIX))
-
-
-@bp.route("/version")
-def version():
-    """
-    Return notebook services version.
-
-    ---
-    get:
-      description: Information about notebooks service.
-      responses:
-        200:
-          description: Notebooks service info.
-          content:
-            application/json:
-              schema: VersionResponse
-    """
-    from ..version import __version__
-
-    info = {
-        "name": "renku-notebooks",
-        "versions": [
-            {
-                "version": __version__,
-                "data": {
-                    "anonymous_sessions_enabled": config.ANONYMOUS_SESSIONS_ENABLED,
-                    "s3_mount_enabled": config.S3_MOUNTS_ENABLED,
-                },
-            }
-        ],
-    }
-    return VersionResponse().dump(info), 200
 
 
 blueprints = [bp, notebooks.bp]
